+++ src/Exception/ParsingExecption.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Exception;
+
+class ParsingExecption extends \Exception
+{
+}