+++ CHANGELOG.md
@@ -10,6 +10,8 @@ See [keep a changelog] for information about writing changes to this log.
 
 ### Added
 
+- Metrics bundlen added.
+- Data model.
 - Symfony core.
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/